+++ pytorch_lightning/trainer/trainer.py
@@ -57,7 +57,6 @@ from pytorch_lightning.plugins import (
 from pytorch_lightning.plugins.environments.slurm_environment import SLURMEnvironment
 from pytorch_lightning.profiler import (
     AdvancedProfiler,
-    BaseProfiler,
     PassThroughProfiler,
     Profiler,
     PyTorchProfiler,
@@ -138,7 +137,6 @@ class Trainer(
         default_root_dir: Optional[str] = None,
         gradient_clip_val: Optional[Union[int, float]] = None,
         gradient_clip_algorithm: Optional[str] = None,
-        process_position: int = 0,
         num_nodes: int = 1,
         num_processes: Optional[int] = None,  # TODO: Remove in 2.0
         devices: Optional[Union[List[int], str, int]] = None,
@@ -173,7 +171,7 @@ class Trainer(
         weights_save_path: Optional[str] = None,  # TODO: Remove in 1.8
         num_sanity_val_steps: int = 2,
         resume_from_checkpoint: Optional[Union[Path, str]] = None,
-        profiler: Optional[Union[BaseProfiler, str]] = None,
+        profiler: Optional[Union[Profiler, str]] = None,
         benchmark: Optional[bool] = None,
         deterministic: Union[bool, _LITERAL_WARN] = False,
         reload_dataloaders_every_n_epochs: int = 0,
@@ -305,13 +303,6 @@ class Trainer(
 
             log_every_n_steps: How often to log within steps.
                 Default: ``50``.
-
-            process_position: Orders the progress bar when running multiple models on same machine.
-
-                .. deprecated:: v1.5
-                    ``process_position`` has been deprecated in v1.5 and will be removed in v1.7.
-                    Please pass :class:`~pytorch_lightning.callbacks.progress.TQDMProgressBar` with ``process_position``
-                    directly to the Trainer's ``callbacks`` argument instead.
 
             enable_progress_bar: Whether to enable to progress bar by default.
                 Default: ``False``.
@@ -509,7 +500,6 @@ class Trainer(
             callbacks,
             enable_checkpointing,
             enable_progress_bar,
-            process_position,
             default_root_dir,
             weights_save_path,
             enable_model_summary,
@@ -792,8 +782,8 @@ class Trainer(
 
         Returns:
             List of dictionaries with metrics logged during the validation phase, e.g., in model- or callback hooks
-            like :meth:`~pytorch_lightning.core.lightning.LightningModule.validation_step`,
-            :meth:`~pytorch_lightning.core.lightning.LightningModule.validation_epoch_end`, etc.
+            like :meth:`~pytorch_lightning.core.module.LightningModule.validation_step`,
+            :meth:`~pytorch_lightning.core.module.LightningModule.validation_epoch_end`, etc.
             The length of the list corresponds to the number of validation dataloaders used.
         """
         self.strategy.model = model or self.lightning_module
@@ -880,8 +870,8 @@ class Trainer(
 
         Returns:
             List of dictionaries with metrics logged during the test phase, e.g., in model- or callback hooks
-            like :meth:`~pytorch_lightning.core.lightning.LightningModule.test_step`,
-            :meth:`~pytorch_lightning.core.lightning.LightningModule.test_epoch_end`, etc.
+            like :meth:`~pytorch_lightning.core.module.LightningModule.test_step`,
+            :meth:`~pytorch_lightning.core.module.LightningModule.test_epoch_end`, etc.
             The length of the list corresponds to the number of test dataloaders used.
         """
         self.strategy.model = model or self.lightning_module